src/backend/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "FastAPI backend for PyReflect interface"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "numpy>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pymongo>=4.6.0",
    "pyreflect @ git+https://github.com/williamQyq/pyreflect.git@main",
    "boto3>=1.42.29",
    "huggingface-hub>=1.3.2",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = []

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cpu"]
environments = ["sys_platform == 'darwin' and platform_machine == 'arm64'", "sys_platform == 'linux'"]
index-strategy = "unsafe-best-match"
